// src/app/interfaces/calendar.interface.ts
// calendar.interface.ts

// Generated by https://quicktype.io

export interface Event {
  id?: number;
  title?: string | null;
  start_date?: string | null;
  end_date?: string | null;
  all_day?: boolean;
  description?: string | null;
  location?: string | null;
  color?: string | null;
  state?: boolean;
}